+++ packages/core/util/util.js
@@ -12,6 +12,8 @@
  * @exports Job
  * @exports memoize
  * @exports mergeClassNameMaps
+ * @exports normalizePublicClassNames
+ * @exports applyDefaultProps
  * @exports perfNow
  * @exports mapAndFilterChildren
  * @exports shallowEqual
@@ -21,7 +23,7 @@ import {checkPropTypes as check} from 'prop-types';
 import always from 'ramda/src/always';
 import isType from 'ramda/src/is';
 import unless from 'ramda/src/unless';
-import {Children, useState} from 'react';
+import {Children, useRef} from 'react';
 import * as ReactIs from 'react-is';
 
 import Job from './Job';
@@ -210,6 +212,58 @@ const mergeClassNameMaps = (baseMap, additiveMap, allowedClassNames) => {
 };
 
 /**
+ * Normalizes a `publicClassNames` config value into an array of logical class names.
+ *
+ * @function
+ * @param {Boolean|String|String[]} [publicClassNames]  The public class names config
+ * @param {Object}                  [css]             Component CSS map used when
+ *                                                      `publicClassNames` is `true`
+ * @returns {String[]|Boolean|String|undefined}         Normalized class name list, or the
+ *                                                      original value when no normalization applies
+ * @memberof core/util
+ * @public
+ */
+const normalizePublicClassNames = (publicClassNames, css) => {
+	let allowedClassNames = publicClassNames;
+
+	if (css && allowedClassNames === true) {
+		allowedClassNames = Object.keys(css);
+	} else if (typeof allowedClassNames === 'string') {
+		allowedClassNames = allowedClassNames.split(/\s+/);
+	}
+
+	return allowedClassNames;
+};
+
+/**
+ * Applies default values for keys that are `undefined` on `target`.
+ *
+ * When `keys` is provided, only those keys are considered. This avoids
+ * `Object.keys` on the hot path (e.g. `core/kind` render).
+ *
+ * @function
+ * @param {Object}        target         Props object to update in place
+ * @param {Object}        defaultProps   Default value object
+ * @param {String[]}      [keys]         Optional precomputed key list
+ *
+ * @returns {Object}                      The updated `target`
+ * @memberof core/util
+ * @public
+ */
+const applyDefaultProps = (target, defaultProps, keys) => {
+	if (keys?.length) {
+		keys.forEach(key => {
+			// eslint-disable-next-line no-undefined
+			if (target[key] === undefined) {
+				target[key] = defaultProps[key];
+			}
+		});
+	}
+
+	return target;
+};
+
+/**
  * Creates a function that memoizes the result of `fn`.
  *
  * Note that this function is a naive implementation and only checks the first argument for
@@ -281,16 +335,7 @@ const mapAndFilterChildren = (children, callback, filter) => {
  * @public
  */
 const setDefaultProps = (props, defaultProps = {}) => {
-	const result = Object.assign({}, props);
-
-	for (const prop in defaultProps) {
-		// eslint-disable-next-line no-undefined
-		if (props[prop] === undefined) {
-			result[prop] = defaultProps[prop];
-		}
-	}
-
-	return result;
+	return applyDefaultProps(Object.assign({}, props), defaultProps, Object.keys(defaultProps));
 };
 
 /**
@@ -373,19 +418,17 @@ const checkPropTypes = (component, props, prevProps) => {
  * @public
  */
 const usePrevious = (value) => {
-	const [previousTrackedValue, setPreviousTrackedValue] = useState(value);
-	const [previousValue, setPreviousValue] = useState(value);
+	const ref = useRef(value);
+	const previous = ref.current;
 
-	if (value !== previousTrackedValue) {
-		setPreviousTrackedValue(value);
-		setPreviousValue(previousTrackedValue);
-	}
+	ref.current = value;
 
-	return previousValue;
+	return previous;
 };
 
 export {
 	cap,
+	applyDefaultProps,
 	checkPropTypes,
 	clamp,
 	coerceArray,
@@ -395,6 +438,7 @@ export {
 	Job,
 	memoize,
 	mergeClassNameMaps,
+	normalizePublicClassNames,
 	perfNow,
 	mapAndFilterChildren,
 	setDefaultProps,